Ian/requirements.txt
gradio
transformers
accelerate
torch
Pillow

